Move message handler to separate file along with retry mechanism
