+++ .env.development
@@ -8,3 +8,5 @@ ETHEREUM_PROVIDER=alchemy
 ETHEREUM_PROVIDER_API_KEY=
 
 MONGODB_SERVER="mongodb://root:root@localhost:27017/admin"
+
+MESSAGE_MAX_RETRY=3
